Fixes #673.
Also, set level "premium" on users from social.
