+++ lib/auth.js
@@ -10,7 +10,8 @@ var express = require('express'),
 	User = require('../models/user.js').User,
 	util = require('util'),
 	logger = require('./logger'),
-	conf = require('config');
+	conf = require('config'),
+	user_utils = require('./user-utils');
 
 passport.serializeUser(function (user, done) {
 	done(null, user._id);
@@ -155,7 +156,10 @@ function addProviderToUser(provider, providerId, profile, currentUser, done) {
 				if (err) {
 					throw err;
 				}
-				done(null, user);
+
+				user_utils.initUser(user, function () {
+					done(null, user);
+				});
 			});
 		}
 	})
