+++ public/index.html
@@ -61,11 +61,11 @@
                     </div>
                 </form>
                 <p id="form-success" style="display:none;"></p>
-				<!--<div class="socialicons clearfix">
+				<div class="socialicons clearfix">
 					<div class="socialicon"><a href="auth/facebook"><span class="icon-facebook"></span></a></div>
 					<div class="socialicon"><a href="auth/linkedin"><span class="icon-linkedin"></span></a></div>
 					<div class="socialicon"><a href="auth/google"><span class="icon-google-plus"></span></a></div>
-				</div>-->
+				</div>
 			 </div>
 		</article>
 	</section>
