+++ routes/user.js
@@ -6,20 +6,13 @@ var User = require('../models/user.js').User
 	, logger = require('../lib/logger')
 	, conf = require('config')
 	, env = process.env.NODE_ENV
-	, path = require('path')
 	, fs = require('fs')
 	, utils = require('../lib/utils')
 	, utils_shared = require('../public/create/scripts/utils-shared')
-	, styleset_utils = require('../lib/styleset-utils')
-	, mkdirp = require('mkdirp')
+	, user_utils = require('../lib/user-utils')
 	, Styleset = require('../models/styleset.js').Styleset
-	, copyStyleset = require('../models/styleset.js').copy
 	, discourse_sso = require('discourse-sso')
 ;
-
-function hashEmail(email) {
-	return crypto.createHash('md5').update(conf.app.salt + email).digest("hex");
-}
 
 /**
  * GET current user.
@@ -154,95 +147,8 @@ exports.register = function (req, res, next) {
 			isDemo: req.body.isDemo
 		});
 
-		if ('test' != env) {
-			// Currently we just force user to become premium member immediately (for free)
-			// TODO: When launching (currently Beta1) this should be removed!
-			user.level = "premium";
-		}
-
-		// Copy all system/Scripler stylesets (and styles) to the user
-		Styleset.find({isSystem: true}, function (err, stylesets) {
-			if (err) {
-				return next(err);
-			}
-
-			user.save(function (err) {
-				if (err) {
-					// return error
-					// TODO: add 11001 as in User.edit()?
-					if (err.code == 11000) {
-						return next({errors: "Email already registered", status: 400});
-					}
-					return next(err);
-				} else {
-					if ('test' != env) {
-						if (!user.isDemo) {
-							emailer.sendUserEmail(
-								user,
-								[
-									{name: "URL", content: conf.app.url_prefix + 'user/' + user._id + '/verify/' + hashEmail(user.email)}
-								],
-								'welcome'
-							);
-						}
-					}
-				}
-
-				var createDirectories = function (next) {
-					var userDir = path.join(conf.resources.usersDir, conf.epub.userDirPrefix + user._id);
-					mkdirp(userDir, function (err) {
-						if (err) {
-							return next(err);
-						} else {
-							return res.send({"user": utils.cleanUserObject(user)});
-						}
-					});
-				};
-
-				var numberOfStylesetsToBeCopied = stylesets.length;
-				if (numberOfStylesetsToBeCopied == 0) {
-					return createDirectories(next);
-				} else {
-					var stylesetCopies = [];
-					stylesets.forEach(function (styleset) {
-						styleset.isSystem = false;
-						styleset.members = [{userId: user._id, access: ["admin"]}];
-						copyStyleset(styleset, function(err, copy) {
-							if (err) {
-								return next(err);
-							}
-
-							stylesetCopies.push(copy);
-
-							if (copy.name === conf.user.defaultStylesetName) {
-								user.defaultStyleset = copy;
-							}
-
-							numberOfStylesetsToBeCopied--;
-
-							if (numberOfStylesetsToBeCopied == 0) {
-								// Sort stylesets by name (currently only used for integration test, i.e. not important for the app)
-								// After stylesets are added to user.stylesets, user.stylesets only contains ids, so to avoid re-reading the stylesets from the db, save them in a temporary array.
-								stylesetCopies.sort(styleset_utils.systemStylesetOrder);
-								user.stylesets = stylesetCopies;
-
-								user.save(function (err) {
-									if (err) {
-										return next(err);
-									}
-
-									if (utils.isEmpty(user.defaultStyleset)) {
-										// TODO: should this error be shown to the user?
-										logger.error("No default styleset set for user " + user.firstname + " " + user.lastname + "(id = " + user._id + ").");
-									}
-
-									return createDirectories(next);
-								});
-							}
-						});
-					});
-				}
-			});
+		user_utils.initUser(user, function () {
+			res.send({"user": utils.cleanUserObject(user)})
 		});
 	}
 };
